+++ src/main/java/net/thedragonteam/armorplus/util/ArmorPlusItemUtils.java
@@ -9,7 +9,7 @@ import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 
 import static net.thedragonteam.armorplus.APConfig.getRD;
-import static net.thedragonteam.thedragonlib.util.ItemStackUtils.getItemStack;
+import static net.thedragonteam.thedragonlib.util.ItemStackUtils.getItem;
 
 /**
  * ArmorPlus created by sokratis12GR
@@ -17,47 +17,39 @@ import static net.thedragonteam.thedragonlib.util.ItemStackUtils.getItemStack;
  */
 public class ArmorPlusItemUtils {
 
-    public static boolean isItemRepairable(ItemStack repair, ItemStack easy, ItemStack expert) {
-        switch (getRD()) {
-            case EASY:
-                return repair == easy;
-            case EXPERT:
-                return repair == expert;
-            case HELLISH:
-                return false;
-        }
-        return false;
+    public static boolean isItemRepairable(ItemStack repair, Item easy, Item expert) {
+        return getRD().isItemRepairable(repair, easy, expert);
     }
 
-    public static boolean isItemRepairable(ItemStack repair, Item easy, Item expert) {
-        return isItemRepairable(repair, getItemStack(easy), getItemStack(expert));
+    public static boolean isItemRepairable(ItemStack repair, ItemStack easy, ItemStack expert) {
+        return isItemRepairable(repair, getItem(easy), getItem(expert));
     }
 
     public static boolean isItemRepairable(ItemStack repair, Block easy, ItemStack expert) {
-        return isItemRepairable(repair, getItemStack(easy), expert);
+        return isItemRepairable(repair, getItem(easy), getItem(expert));
     }
 
     public static boolean isItemRepairable(ItemStack repair, Block easy, Block expert) {
-        return isItemRepairable(repair, getItemStack(easy), getItemStack(expert));
+        return isItemRepairable(repair, getItem(easy), getItem(expert));
     }
 
     public static boolean isItemRepairable(ItemStack repair, ItemStack easy, Block expert) {
-        return isItemRepairable(repair, easy, getItemStack(expert));
+        return isItemRepairable(repair, getItem(easy), getItem(expert));
     }
 
     public static boolean isItemRepairable(ItemStack repair, ItemStack easy, Item expert) {
-        return isItemRepairable(repair, easy, getItemStack(expert));
+        return isItemRepairable(repair, getItem(easy), expert);
     }
 
     public static boolean isItemRepairable(ItemStack repair, Item easy, ItemStack expert) {
-        return isItemRepairable(repair, getItemStack(easy), expert);
+        return isItemRepairable(repair, easy, getItem(expert));
     }
 
     public static boolean isItemRepairable(ItemStack repair, Block easy, Item expert) {
-        return isItemRepairable(repair, getItemStack(easy), getItemStack(expert));
+        return isItemRepairable(repair, getItem(easy), expert);
     }
 
     public static boolean isItemRepairable(ItemStack repair, Item easy, Block expert) {
-        return isItemRepairable(repair, getItemStack(easy), getItemStack(expert));
+        return isItemRepairable(repair, easy, getItem(expert));
     }
 }